% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{check_hz}
\alias{check_hz}
\title{Checks whether we are importing something that is measured in hz. If it is, it changes the row names to reflect the frequency}
\usage{
check_hz(data)
}
\arguments{
\item{data}{The dataframe we want to check. If}
}
\description{
Checks whether we are importing something that is measured in hz. If it is, it changes the row names to reflect the frequency
}
