% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{fix_hz_index}
\alias{fix_hz_index}
\title{Small helper function that converts index into hz}
\usage{
fix_hz_index(x, freq)
}
\arguments{
\item{x}{a string with an integer}

\item{freq}{the frequency of your dataset}
}
\description{
Small helper function that converts index into hz
}
