% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{load_case}
\alias{load_case}
\title{PENDING DESCRIPTION OF FUNCTION}
\usage{
load_case(tnames, caseid)
}
\arguments{
\item{tnames}{PENDING EXPLANATION OF VARIABLE}

\item{caseid}{PENDING EXPLANATION OF VARIABLE}
}
\description{
PENDING DESCRIPTION OF FUNCTION
}
