% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{load_trk}
\alias{load_trk}
\title{Loads a data track from VitalDB api}
\usage{
load_trk(tid)
}
\arguments{
\item{tid}{PENDING EXPLANATION OF VARIABLE}
}
\description{
Loads a data track from VitalDB api
}
