% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{load_VDB}
\alias{load_VDB}
\title{Loads file from URL (helper function)}
\usage{
load_VDB(file_url)
}
\arguments{
\item{file_url}{URL for API endpoint}
}
\description{
Loads file from URL (helper function)
}
