% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{plot_art}
\alias{plot_art}
\title{Function for plotting ART data}
\usage{
plot_art(art_data, insp_start_data, beats_data)
}
\arguments{
\item{art_data}{Dataframe with ART data}

\item{insp_start_data}{Inspiration start data}

\item{beats_data}{Beats data}
}
\description{
Function for plotting ART data
}
