% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{plot_awp}
\alias{plot_awp}
\title{Function for plotting AWP data}
\usage{
plot_awp(awp_data, add_insp_start = "no", insp_start_data = NULL)
}
\arguments{
\item{awp_data}{Dataframe with AWP data}
}
\description{
Function for plotting AWP data
}
