% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{plot_convolution_data}
\alias{plot_convolution_data}
\title{Function for plotting Convolution Data}
\usage{
plot_convolution_data(convolution_data)
}
\arguments{
\item{convolution_data}{Dataframe with convolution_data}
}
\description{
Function for plotting Convolution Data
}
