% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{pp_plot}
\alias{pp_plot}
\title{Function for plotting Pulse Pressure}
\usage{
pp_plot(insp_start_data, beats_data)
}
\arguments{
\item{insp_start_data}{Inspiration start data}

\item{beats_data}{Beats data}
}
\description{
Function for plotting Pulse Pressure
}
