% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{pp_plot_color}
\alias{pp_plot_color}
\title{Function for plotting Pulse Pressure where different respiratory cycles are indicated}
\usage{
pp_plot_color(beats_indexed_data, insp_start_data)
}
\description{
Function for plotting Pulse Pressure where different respiratory cycles are indicated
}
