% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{pp_plot_color_and_index}
\alias{pp_plot_color_and_index}
\title{Function for plotting pp_plot_insp and pp_plot_color}
\usage{
pp_plot_color_and_index(beats_indexed_data, insp_start_data)
}
\description{
Function for plotting pp_plot_insp and pp_plot_color
}
