% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{pp_plott_color}
\alias{pp_plott_color}
\title{Function for plotting Pulse Pressure where different respiratory cycles are indicated}
\usage{
pp_plott_color(beats_indexed_data, insp_start_data)
}
\arguments{
\item{beats_indexed_data}{Beats with indexes data}

\item{insp_start_data}{Inspiration start data}
}
\description{
Function for plotting Pulse Pressure where different respiratory cycles are indicated
}
