% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{get_inspiration_start}
\alias{get_inspiration_start}
\title{Function for finding inspiration starts}
\usage{
get_inspiration_start(df)
}
\arguments{
\item{df}{Dataframe with AWP}
}
\description{
Function for finding inspiration starts
}
