% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{load_case}
\alias{load_case}
\title{Main function for loading in data}
\usage{
load_case(tname, caseid)
}
\arguments{
\item{tname}{Name of track}

\item{caseid}{A case id, from which you want to load the track given as tname}
}
\description{
Main function for loading in data
}
