% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{load_trk}
\alias{load_trk}
\title{Loads a data track from VitalDB api (helper function)}
\usage{
load_trk(tid)
}
\arguments{
\item{tid}{Track id}
}
\description{
Loads a data track from VitalDB api (helper function)
}
