% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalDBR.R
\name{load_VDB}
\alias{load_VDB}
\title{Loads file from URL (helper function)}
\usage{
load_VDB(file_url)
}
\arguments{
\item{file_url}{URL for API endpoint. Found at https://vitaldb.net/dataset/?query=api&documentId=1_VTteoI5v-cdFkUTSnhYoGCqyq8dAC-_k50oM1QOjkU&sectionId=h.av1x1sa5y1he}
}
\value{
A dataframe with values from the specified endpoint
}
\description{
Loads file from URL (helper function)
}
